move micro benchmarks to the directory benches
