+++ benches/micro/Cargo.toml
@@ -1,19 +1,24 @@
-[workspace]
-members = [
-    "hermit-abi",
-    "hermit-sys",
-    "benches/netbench",
-    "benches/micro",
-    "examples/httpd",
-    "examples/demo",
-]
-exclude = ["target", "loader", "libhermit-rs"]
+[package]
+name = "micro_benchmarks"
+version = "0.1.0"
+authors = ["Stefan Lankes <slankes@eonerc.rwth-aachen.de>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+rayon = "1.3.0"
+
+[target.'cfg(all(any(target_arch = "x86_64", target_arch = "aarch64"), target_os = "hermit"))'.dependencies]
+hermit-sys = { path = "../../hermit-sys", default-features = false, features = ["with_submodule"] }
+
+[target.'cfg(target_os = "linux")'.dependencies]
+syscalls = { version = "0.2", default-features = false }
 
 [profile.release]
 opt-level = 3
 debug = false
 rpath = false
-lto = "thin"
+lto = true
 debug-assertions = false
 
 [profile.dev]
